refactor：优化代码框架
