+++ apis_flask.py
@@ -1,164 +1,33 @@
-import copy
 import hashlib
 import json
 import logging
 import os
-import re
 import subprocess
+import traceback
 
 from datetime import datetime, timedelta
-from functools import wraps
 from urllib.parse import urlparse
 
 # 第三方库导入 (Third-party packages)
-import requests
-from bs4 import BeautifulSoup
-from flask import Flask, abort, flash, redirect, render_template, request, session, url_for
 
 # Django相关导入 (Django framework)
-import django
 from django.db import IntegrityError
 from django.db.models import Q
-from flask import Flask, render_template, request, redirect, url_for, flash, session, abort
+from flask import Flask, render_template, request, redirect, url_for, flash, session
 from django.db import transaction
 
-from django_config import configure_django
-
-from django.utils import timezone
-
+from common import fetch_title
 # 本地应用导入 (Local application imports)
 from common.decorator.allowed_file import allowed_file
 from common.decorator.is_valid_url import is_valid_url
-from django_config import configure_django
-
-import logging
-import traceback
-
-
-
-
-def get_timezone_aware_datetime(date_str):
-    """将字符串日期转换为时区感知的datetime对象"""
-    naive_dt = datetime.strptime(date_str, '%Y-%m-%d')
-    return timezone.make_aware(naive_dt)
-
-
-# 配置日志系统
-logging.basicConfig(
-    level=logging.DEBUG,
-    format='%(levelname)s - %(asctime)s - %(name)s - [%(filename)s:%(lineno)d] - %(message)s',
-    handlers=[
-        logging.FileHandler('app.log'),
-        logging.StreamHandler()
-    ]
-)
-
-configure_django()
-django.setup()
-
+from decorators import login_required, editor_required
 from django_models.models import User_info, Content
+from global_static import *
+
 
 app = Flask(__name__)
 app.secret_key = 'A_SECRET_KEY_HERE'
 app.permanent_session_lifetime = timedelta(days=7)
-
-EMAIL_ADDRESS = "test@smail.nju.edu.cn"
-EMAIL_PASSWORD = "A_SECRET_KEY_HERE"
-SMTP_SERVER = "smtp.exmail.qq.com"
-SMTP_PORT = 465
-
-FILE_PATH = 'static/uploads'
-
-def fetch_title(url):
-    """从URL获取网页标题"""
-    for _ in range(2):
-        try:
-            headers = {
-                'User-Agent': 'Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/91.0.4472.124 Safari/537.36'
-            }
-            with requests.get(url, headers=headers, stream=True, timeout=5) as response:
-                response.raise_for_status()
-                partial_content = ""
-                for chunk in response.iter_content(chunk_size=4096):
-                    partial_content += chunk.decode('utf-8', errors='replace')
-                    if "</head>" in partial_content.lower():
-                        soup = BeautifulSoup(partial_content, 'html.parser')
-                        meta_tag = soup.find("meta", property="og:title")
-                        if meta_tag and meta_tag.get("content"):
-                            return meta_tag.get("content")
-                        title_tag = soup.find("title")
-                        if title_tag and title_tag.string:
-                            return title_tag.string.strip()
-            soup = BeautifulSoup(partial_content, 'html.parser')
-            meta_tag = soup.find("meta", property="og:title")
-            if meta_tag and meta_tag.get("content"):
-                return meta_tag.get("content")
-        except Exception as e:
-            logging.warning(f"获取标题失败，URL: {url}, 错误: {str(e)}")
-    return "标题获取失败"
-
-
-EDITOR_LIST = ["editor1", "editor2", "111"]
-ADMIN_LIST = ["admin", "222", "111"]
-
-LINK_REGEX = re.compile(
-    r"(https?:\/\/(?:www\.)?[-a-zA-Z0-9@:%._\+~#=]{1,256}\.[a-zA-Z0-9()]{1,6}\b(?:[-a-zA-Z0-9()@:%_\+.~#?&//=]*)?)")
-
-
-
-def login_required(f):
-    """装饰器：要求用户登录"""
-
-    @wraps(f)
-    def decorated_function(*args, **kwargs):
-        if 'username' not in session:
-            return redirect(url_for('login'))
-        return f(*args, **kwargs)
-
-    return decorated_function
-
-
-def admin_required(f):
-    """装饰器：要求管理员权限"""
-
-    @wraps(f)
-    def decorated_function(*args, **kwargs):
-
-        username = session.get('username')
-
-        if not username:
-            abort(401)
-
-        user = User_info.objects.filter(username=username).first()
-
-        if not user or not user.has_admin_permission():
-            abort(403)
-
-        return f(*args, **kwargs)
-
-    return decorated_function
-
-
-def editor_required(f):
-    """装饰器：要求编辑权限"""
-
-    @wraps(f)
-    def decorated_function(*args, **kwargs):
-
-        username = session.get('username')
-
-        if not username:
-            abort(401)
-
-        user = User_info.objects.filter(username=username).first()
-
-        # 如果用户不存在或没有权限，都返回403
-        if not user or not user.has_editor_permission():
-            abort(403)
-
-        return f(*args, **kwargs)
-
-    return decorated_function
 
 
 @app.route('/login', methods=['GET', 'POST'])
@@ -338,38 +207,6 @@ def upload():
         return redirect(url_for('main'))
 
     return render_template('upload.html')
-
-
-# @app.route('/describe/<int:entry_id>', methods=['GET', 'POST'])
-# @login_required
-# def describe(entry_id):
-#     if request.method == 'POST':
-#         title = request.form['title']
-#         entry_type = request.form['entry_type']
-#         description = request.form['description']
-#         due_time = request.form['due_time']
-#         use_image = 1 if request.form.get('use_image') == 'on' else 0
-#         tag = request.form.get('tag')
-#         short_title = request.form.get('short_title') or title
-#         content = Content.objects.create(
-#             uploader=session['username'],
-#             describer=session['username'],
-#             title=title,
-#             short_title=short_title,
-#             content=description,
-#             status='pending',
-#             type=entry_type,
-#             tag=tag,
-#             deadline=due_time
-#         )
-#         return redirect(url_for('main'))
-#     else:
-#         try:
-#             entry = Content.objects.get(id=entry_id)
-#         except Content.DoesNotExist:
-#             entry = None
-#
-#         return render_template('describe.html', entry=entry)
 
 
 @app.route('/describe/<int:entry_id>', methods=['GET', 'POST'])
@@ -571,35 +408,6 @@ def logout():
     return redirect(url_for('login'))
 
 
-#
-# @app.route('/change_password', methods=['GET', 'POST'])
-# @login_required
-# def change_password():
-#     if request.method == 'POST':
-#         original_password = request.form.get('original_password')
-#         new_password = request.form.get('new_password')
-#         confirm_password = request.form.get('confirm_password')
-#         if new_password != confirm_password:
-#             flash("New passwords do not match")
-#             return redirect(url_for('change_password'))
-#         conn = sqlite3.connect('database.db')
-#         c = conn.cursor()
-#         c.execute("SELECT * FROM users WHERE username=?", (session['username'],))
-#         user = c.fetchone()
-#         if user and check_password_hash(user[2], original_password):
-#             c.execute("UPDATE users SET password=? WHERE username=?", (generate_password_hash(new_password), session['username']))
-#             conn.commit()
-#             flash("Password successfully updated. Please log in again.")
-#             session.pop('username', None)
-#             conn.close()
-#             return redirect(url_for('login'))
-#         else:
-#             flash("Original password is incorrect")
-#             conn.close()
-#         return redirect(url_for('change_password'))
-#     return render_template('change_password.html')
-#
-
 @app.route('/paste', methods=['POST'])
 @login_required
 def paste():
@@ -699,192 +507,6 @@ def upload_image():
         flash("不支持的文件格式")
         return redirect(url_for('main'))
 
-
-
-#
-# @app.route('/admin')
-# @editor_required
-# def admin():
-#     page = request.args.get('page', default=1, type=int)
-#     page_size = 12
-#     offset = (page - 1) * page_size
-#     conn = sqlite3.connect('database.db')
-#     c = conn.cursor()
-#     c.execute("SELECT COUNT(*) FROM entries")
-#     total_count = c.fetchone()[0]
-#     total_pages = (total_count + page_size - 1) // page_size
-#     c.execute("SELECT * FROM entries ORDER BY upload_time DESC LIMIT ? OFFSET ?", (page_size, offset))
-#     entries = c.fetchall()
-#     conn.close()
-#     default_publish_date = datetime.now().strftime("%Y-%m-%d")
-#     return render_template('admin.html', entries=entries, page=page, total_pages=total_pages, default_publish_date=default_publish_date)
-#
-# @app.route('/admin/edit/<int:entry_id>', methods=['GET', 'POST'])
-# @admin_required
-# def admin_edit(entry_id):
-#     conn = sqlite3.connect('database.db')
-#     c = conn.cursor()
-#     if request.method == 'POST':
-#         uploader = request.form['uploader']
-#         upload_time = request.form['upload_time']
-#         title = request.form['title']
-#         link = request.form['link']
-#         description = request.form['description']
-#         describer = request.form['describer']
-#         reviewer = request.form['reviewer']
-#         due_time = request.form['due_time']
-#         status = request.form['status']
-#         entry_type = request.form['entry_type']
-#         c.execute("""UPDATE entries SET uploader=?, upload_time=?, title=?, link=?, description=?, describer=?, reviewer=?, due_time=?, status=?, type=? WHERE id=?""",
-#                   (uploader, upload_time, title, link, description, describer, reviewer, due_time, status, entry_type, entry_id))
-#         conn.commit()
-#         conn.close()
-#         flash("条目已更新")
-#         return redirect(url_for('admin'))
-#     else:
-#         c.execute("SELECT * FROM entries WHERE id=?", (entry_id,))
-#         entry = c.fetchone()
-#         conn.close()
-#         return render_template('admin_edit.html', entry=entry)
-#
-# @app.route('/admin/delete/<int:entry_id>', methods=['POST'])
-# @admin_required
-# def admin_delete(entry_id):
-#     conn = sqlite3.connect('database.db')
-#     c = conn.cursor()
-#     c.execute("DELETE FROM entries WHERE id=?", (entry_id,))
-#     conn.commit()
-#     conn.close()
-#     flash("条目已删除")
-#     return redirect(url_for('admin'))
-#
-# @app.route('/admin/publish', methods=['POST'])
-# @editor_required
-# def admin_publish():
-#     entry_ids = request.form.getlist("entry_ids")
-#     publish_time_input = request.form.get("publish_time")
-#     if not entry_ids:
-#         flash("请至少选择一个条目")
-#         return redirect(url_for('admin'))
-#     try:
-#         publish_time = datetime.strptime(publish_time_input, "%Y-%m-%d")
-#     except (ValueError, TypeError):
-#         publish_time = datetime.now()
-#     conn = sqlite3.connect('database.db')
-#     c = conn.cursor()
-#     for eid in entry_ids:
-#         c.execute("UPDATE entries SET publish_date=? WHERE id=?", (publish_time, eid))
-#     conn.commit()
-#     conn.close()
-#     flash("选定条目的发布时间已更新")
-#     return redirect(url_for('admin'))
-#
-# @app.route('/admin/publish_today', methods=['POST'])
-# @editor_required
-# def admin_publish_today():
-#     publish_time = datetime.now().date()
-#     today = datetime.now().date()
-#     conn = sqlite3.connect('database.db')
-#     c = conn.cursor()
-#     c.execute("UPDATE entries SET publish_date=? WHERE date(upload_time)=?", (publish_time, today))
-#     conn.commit()
-#     conn.close()
-#     flash("今天上传的条目的发布时间已更新")
-#     return redirect(url_for('admin'))
-#
-# @app.route('/admin/unpublish', methods=['POST'])
-# @editor_required
-# def admin_unpublish():
-#     entry_ids = request.form.getlist("entry_ids")
-#     if not entry_ids:
-#         flash("请至少选择一个条目")
-#         return redirect(url_for('admin'))
-#     conn = sqlite3.connect('database.db')
-#     c = conn.cursor()
-#     for eid in entry_ids:
-#         c.execute("UPDATE entries SET publish_date=NULL WHERE id=?", (eid,))
-#     conn.commit()
-#     conn.close()
-#     flash("选定条目已取消发布")
-#     return redirect(url_for('admin'))
-#
-# @app.route('/admin/make_invalid', methods=['POST'])
-# @editor_required
-# def admin_make_invalid():
-#     entry_ids = request.form.getlist("entry_ids")
-#     if not entry_ids:
-#         flash("请至少选择一个条目")
-#         return redirect(url_for('admin'))
-#     invalid_date = datetime(1970, 1, 1)
-#     conn = sqlite3.connect('database.db')
-#     c = conn.cursor()
-#     for eid in entry_ids:
-#         c.execute("UPDATE entries SET publish_date=? WHERE id=?", (invalid_date, eid))
-#     conn.commit()
-#     conn.close()
-#     flash("选定条目已标记为无效")
-#     return redirect(url_for('admin'))
-#
-# @app.route('/admin/user_admin', methods=['GET'])
-# @admin_required
-# def user_admin():
-#     conn = sqlite3.connect('database.db')
-#     c = conn.cursor()
-#     c.execute("SELECT id, username FROM users")
-#     users = c.fetchall()
-#     conn.close()
-#     return render_template('user_admin.html', users=users)
-#
-# @app.route('/admin/user_edit/<int:user_id>', methods=['GET', 'POST'])
-# @admin_required
-# def user_edit(user_id):
-#     conn = sqlite3.connect('database.db')
-#     c = conn.cursor()
-#     c.execute("SELECT id, username FROM users WHERE id=?", (user_id,))
-#     user = c.fetchone()
-#     if not user:
-#         conn.close()
-#         flash("User not found")
-#         return redirect(url_for('user_admin'))
-#
-#     if request.method == 'POST':
-#         new_password = request.form.get('new_password')
-#         if not new_password:
-#             flash("请输入新密码")
-#             return redirect(url_for('user_edit', user_id=user_id))
-#         c.execute("UPDATE users SET password=? WHERE id=?", (generate_password_hash(new_password), user_id))
-#         conn.commit()
-#         conn.close()
-#         flash("密码更新成功")
-#         return redirect(url_for('user_admin'))
-#     conn.close()
-#     return render_template('user_edit.html', user=user)
-#
-# @app.route('/search', methods=['GET'])
-# @login_required
-# def search():
-#     page = request.args.get('page', default=1, type=int)
-#     page_size = 5
-#     offset = (page - 1) * page_size
-#     query = request.args.get('q', '').strip()
-#     results = []
-#     total_pages = 0
-#
-#     if query:
-#         like_query = query  # 不需要手动添加 %%，Django ORM 的 __icontains 会自动处理
-#
-#         # 获取总数
-#         total_count = Content.objects.filter(
-#             Q(title__icontains=query) | Q(description__icontains=query)
-#         ).count()
-#         total_pages = (total_count + page_size - 1) // page_size
-#
-#         # 获取分页结果
-#         results = Content.objects.filter(
-#             Q(title__icontains=query) | Q(description__icontains=query)
-#         ).order_by('created_at')[offset:offset + page_size]
-#
-#     return render_template('search.html', query=query, results=results, page=page, total_pages=total_pages)
 
 @app.route('/search', methods=['GET'])
 @login_required
@@ -1113,152 +735,6 @@ def delete_entry(entry_id):
         return redirect(url_for('login'))
 
 
-#
-# @app.route('/message_board', methods=['GET'])
-# @login_required
-# def message_board():
-#     conn = sqlite3.connect('database.db')
-#     conn.row_factory = sqlite3.Row
-#     c = conn.cursor()
-#     c.execute("SELECT * FROM messages WHERE reply_to IS NULL ORDER BY timestamp DESC")
-#     messages = [dict(row) for row in c.fetchall()]
-#     for message in messages:
-#         c.execute("SELECT * FROM messages WHERE reply_to = ? ORDER BY timestamp ASC", (message['id'],))
-#         message['replies'] = [dict(row) for row in c.fetchall()]
-#     conn.close()
-#     return render_template('message_board.html', messages=messages)
-#
-# @app.route('/post_message', methods=['POST'])
-# @login_required
-# def post_message():
-#     content = request.form['content']
-#     reply_to = request.form.get('reply_to') or None
-#     conn = sqlite3.connect('database.db')
-#     c = conn.cursor()
-#     c.execute("INSERT INTO messages (author, content, reply_to) VALUES (?, ?, ?)",
-#               (session['username'], content, reply_to))
-#     conn.commit()
-#     conn.close()
-#     flash("留言已发表")
-#     return redirect(url_for('message_board'))
-#
-# @app.route('/reply/<int:message_id>', methods=['GET'])
-# @login_required
-# def reply_message(message_id):
-#     conn = sqlite3.connect('database.db')
-#     conn.row_factory = sqlite3.Row
-#     c = conn.cursor()
-#     c.execute("SELECT * FROM messages WHERE id=?", (message_id,))
-#     orig = c.fetchone()
-#     conn.close()
-#     return render_template('reply.html', orig=orig)
-#
-# @app.route('/pastebin', methods=['GET', 'POST'])
-# # @login_required
-# def pastebin():
-#     if request.method == 'POST':
-#         title = request.form.get('title', '').strip()
-#         content = request.form.get('content', '').strip()
-#         if not content:
-#             flash("Paste content cannot be empty")
-#             return redirect(url_for('pastebin'))
-#         try:
-#             uname = session['username']
-#         except:
-#             uname = 'guest'
-#         hash_input = title + content + uname + str(datetime.now())
-#         digest = hashlib.sha256(hash_input.encode('utf-8')).digest()
-#         paste_hash = base64.urlsafe_b64encode(digest).decode('utf-8').rstrip("=")[:10]
-#         conn = sqlite3.connect('database.db')
-#         c = conn.cursor()
-#         c.execute("INSERT INTO pastes (hash, title, content, uploader) VALUES (?, ?, ?, ?)",
-#                   (paste_hash, title, content, uname))
-#         conn.commit()
-#         conn.close()
-#         flash("Paste created successfully")
-#         return redirect(url_for('pastebin_view', paste_hash=paste_hash))
-#     return render_template('pastebin.html')
-#
-# @app.route('/pastebin/<paste_hash>')
-# def pastebin_view(paste_hash):
-#     conn = sqlite3.connect('database.db')
-#     c = conn.cursor()
-#     c.execute("SELECT title, content, uploader, upload_time FROM pastes WHERE hash=?", (paste_hash,))
-#     paste = c.fetchone()
-#     conn.close()
-#     if not paste:
-#         flash("Paste not found")
-#         return redirect(url_for('paste_bin'))
-#     return render_template('pastebin_view.html', paste=paste)
-#
-#
-# @app.route('/admin/files', methods=['GET', 'POST'])
-# @admin_required
-# def admin_files():
-#     upload_folder = os.path.join(app.root_path, 'static', 'admin_uploads')
-#     os.makedirs(upload_folder, exist_ok=True)
-#
-#     if request.method == 'POST':
-#         if 'uploadFile' not in request.files:
-#             flash("没有文件上传")
-#             return redirect(url_for('admin_files'))
-#         file = request.files['uploadFile']
-#         if file.filename == '':
-#             flash("未选择文件")
-#             return redirect(url_for('admin_files'))
-#         if file:
-#             filename = file.filename
-#             file_path = os.path.join(upload_folder, filename)
-#             file.save(file_path)
-#             flash("文件上传成功")
-#             return redirect(url_for('admin_files'))
-#         else:
-#             flash("只允许上传 .html 和 .pdf 文件")
-#             return redirect(url_for('admin_files'))
-#     files = os.listdir(upload_folder)
-#     return render_template('admin_files.html', files=files)
-#
-# @app.route('/admin/files/delete/<filename>', methods=['POST'])
-# @admin_required
-# def admin_file_delete(filename):
-#     upload_folder = os.path.join(app.root_path, 'static', 'admin_uploads')
-#     file_path = os.path.join(upload_folder, filename)
-#     if os.path.exists(file_path):
-#         os.remove(file_path)
-#         flash("文件已删除")
-#     else:
-#         flash("文件未找到")
-#     return redirect(url_for('admin_files'))
-#
-# @app.route('/add_deadline', methods=['GET', 'POST'])
-# @login_required
-# def add_deadline():
-#     if request.method == 'POST':
-#         link = request.form.get('link', '').strip()
-#         link_value = link if link else None
-#         short_title = request.form.get('short_title', '').strip()
-#         tag = request.form.get('tag', '').strip()
-#         today = datetime.now().strftime("%Y-%m-%d")
-#         publish_time = request.form.get('publish_time', today)
-#         due_time = request.form.get('due_time', today)
-#         news = Content.objects.create(
-#             creator_id=session['username'],
-#             describer_id=session['username'],
-#             title=short_title,
-#             link=link_value,
-#             short_title=short_title,
-#             content='',
-#             deadline=due_time,
-#             publish_at=publish_time,
-#             status='pending',
-#             tag=tag,
-#             type="DDLOnly",
-#         )
-#         flash("Deadline entry added successfully")
-#         return redirect(url_for('main'))
-#     today = datetime.now().strftime("%Y-%m-%d")
-#     return render_template('add_deadline.html', today=today)
-
 @app.route('/add_deadline', methods=['GET', 'POST'])
 @login_required
 def add_deadline():
@@ -1324,169 +800,6 @@ def publish():
             with open("latest.json", "r") as f:
                 content = f.read()
         return render_template("publish.html", content=content)
-
-
-#
-# def read_mailing_list():
-#     try:
-#         with open('mailinglist', 'r') as f:
-#             emails = [line.strip() for line in f.readlines() if line.strip()]
-#             return emails
-#     except FileNotFoundError:
-#         return []
-#
-# def generate_news_html(date):
-#     news_data = typst(date)
-#
-#     toc_html = "<h2>目录</h2><ul>"
-#     toc_counter = 0
-#
-#     sections = {
-#         'other': '校级活动',
-#         'lecture': '讲座',
-#         'college': '院级活动',
-#         'club': '社团活动'
-#     }
-#
-#     for section_key, section_title in sections.items():
-#         has_news = section_key in news_data['data'] and news_data['data'][section_key]
-#         has_due = section_key in news_data['due'] and news_data['due'][section_key]
-#
-#         if has_news or has_due:
-#             toc_counter += 1
-#             section_id = f"section-{toc_counter}"
-#             toc_html += f'<li><a href="#{section_id}">{section_title}</a>'
-#
-#             if has_news:
-#                 toc_html += '<ul>'
-#                 for i, item in enumerate(news_data['data'][section_key]):
-#                     item_id = f"{section_id}-item-{i+1}"
-#                     toc_html += f'<li><a href="#{item_id}">{item["title"]}</a></li>'
-#                 toc_html += '</ul>'
-#
-#             toc_html += '</li>'
-#
-#     toc_html += "</ul><hr>"
-#
-#     html_content = f"""
-#     <html>
-#     <head>
-#         <meta charset="UTF-8">
-#         <link rel="preload" href="https://nik-nul.github.io/css/news.css" as="style" onload="this.onload=null;this.rel='stylesheet'">
-#         <noscript><link rel="stylesheet" href="https://nik-nul.github.io/css/news.css"></noscript>
-#         <title>南哪消息 {date}</title>
-#     </head>
-#     <body>
-#         <h1>南哪消息 {date}</h1>
-#         <div class="toc">
-#             {toc_html}
-#         </div>
-#     """
-#
-#     toc_counter = 0
-#     for section_key, section_title in sections.items():
-#         has_news = section_key in news_data['data'] and news_data['data'][section_key]
-#         has_due = section_key in news_data['due'] and news_data['due'][section_key]
-#
-#         if has_news or has_due:
-#             toc_counter += 1
-#             section_id = f"section-{toc_counter}"
-#             html_content += f'<div class="section"><h2 id="{section_id}">{section_title}</h2>'
-#
-#             if has_due:
-#                 html_content += '<table><thead><tr><th>活动标题</th><th>截止时间</th><th>刊载时间</th></tr></thead><tbody>'
-#
-#                 for item in news_data['due'][section_key]:
-#                     link_html = f'<a href="{item["link"]}" target="_blank">{item["title"]}</a>' if item.get('link') else item['title']
-#                     due_time = item['due_time'][:10] if len(item['due_time']) >= 10 else item['due_time']
-#                     publish_time = item['publish_date'][:10] if item.get('publish_date') and len(item['publish_date']) >= 10 else item.get('publish_date', '')
-#                     html_content += f'<tr><td>{link_html}</td><td>{due_time}</td><td>{publish_time}</td></tr>'
-#
-#                 html_content += '</tbody></table>'
-#
-#             if has_news:
-#                 html_content += '<ul>'
-#
-#                 for i, item in enumerate(news_data['data'][section_key]):
-#                     item_id = f"{section_id}-item-{i+1}"
-#                     html_content += f'<li><h3 id="{item_id}">{item["title"]}</h3>'
-#
-#                     if item.get('link'):
-#                         html_content += f'原文链接： <a href="{item["link"]}" target="_blank">{item["link"]}</a>'
-#
-#                     if item.get('description'):
-#                         desc_html = ""
-#                         for desc_part in item['description']:
-#                             if desc_part['type'] == 'text':
-#                                 desc_html += desc_part['content'].replace('\n', '<br>')
-#                             elif desc_part['type'] == 'link':
-#                                 desc_html += f' <a href="{desc_part["content"]}" target="_blank">{desc_part["content"]}</a>'
-#                         html_content += f'<br>{desc_html}'
-#
-#                     html_content += '</li>'
-#
-#                 html_content += '</ul>'
-#
-#             html_content += '</div>'
-#
-#     html_content += """
-#         <hr>
-#         <p>南哪小报编辑部<br>{}</p>
-#         <hr>
-#     </body>
-#     </html>
-#     """.format(date)
-#
-#     return html_content
-#
-# def send_news_email(date, recipient_list=None):
-#     if recipient_list is None:
-#         recipient_list = read_mailing_list()
-#
-#     if not recipient_list:
-#         return False, "邮件列表为空"
-#
-#     html_content = generate_news_html(date)
-#
-#     msg = email.message.EmailMessage()
-#     msg["Subject"] = f"南哪消息 {date}"
-#     msg["From"] = "南哪小报编辑部 <231220103@smail.nju.edu.cn>"
-#     msg["Bcc"] = ", ".join(recipient_list)
-#     msg.set_content(html_content, subtype="html")
-#
-#     for attempt in range(5):
-#         try:
-#             with smtplib.SMTP_SSL(SMTP_SERVER, SMTP_PORT) as server:
-#                 server.login(EMAIL_ADDRESS, EMAIL_PASSWORD)
-#                 server.send_message(msg)
-#             return True, f"邮件发送成功，发送给 {len(recipient_list)} 个收件人"
-#         except Exception as e:
-#             if attempt == 4:
-#                 return False, f"邮件发送失败: {str(e)}"
-#             time.sleep(1)
-#
-#     return False, "邮件发送失败"
-#
-# @app.route('/send_email', methods=['GET', 'POST'])
-# @admin_required
-# def send_email():
-#     if request.method == 'POST':
-#         date = request.form.get('date')
-#         if not date:
-#             flash("请选择日期")
-#             return redirect(url_for('send_email'))
-#
-#         success, message = send_news_email(date)
-#         flash(message)
-#
-#         if success:
-#             return redirect(url_for('send_email'))
-#         else:
-#             return redirect(url_for('send_email'))
-#
-#     today = datetime.now().strftime("%Y-%m-%d")
-#     mailing_list = read_mailing_list()
-#     return render_template('send_email.html', today=today, mailing_list=mailing_list)
 
 
 if __name__ == '__main__':
